* origin/topic/timw/2574-windows-build-warnings:
  Include stdint.h early in flex files to prevent redefinition of int type macros
  Update bifcl submodule [nomail]
  Update binpac submodule [nomail]
  SMB: clarify a confusing line in one of the pac files
  module_util: Change streq helper method to just return bool
  NTLM analyzer: Don't bother checking boolean values for > 0
  FTP analyzer: Remove unnecessary lines from constructor
  IP_Hdr: take length value as uint64_t instead of int
  EventTrace: take size_t as a constructor argument instead of int
  Change argument to Tempvar constructor to size_t, avoids conversion warning in template construction

+++ src/EventTrace.h
@@ -308,7 +308,7 @@ public:
 	// Constructed in terms of the associated script function, "network
 	// time" when the event occurred, and the position of this event
 	// within all of those being traced.
-	EventTrace(const ScriptFunc* _ev, double _nt, int event_num);
+	EventTrace(const ScriptFunc* _ev, double _nt, size_t event_num);
 
 	// Sets a string representation of the arguments (values) being
 	// passed to the event.
